@@ -12,8 +12,8 @@ buildscript {
 
 plugins {
   id 'pl.allegro.tech.build.axion-release' version '1.13.3'
-  id 'org.springframework.boot' version '2.6.14' apply false
-  id 'io.spring.dependency-management' version '1.0.11.RELEASE'
+  id 'org.springframework.boot' version '3.1.5' apply false
+  id 'io.spring.dependency-management' version '1.1.3'
   id 'idea'
   id 'java'
   id 'java-library'
@@ -43,8 +43,8 @@ scmVersion {
 
 idea {
   project {
-    jdkName = "11"
-    languageLevel = "11"
+    jdkName = "17"
+    languageLevel = "17"
   }
 }
 
@@ -58,7 +58,7 @@ subprojects {
 
   group = 'zone.cogni.lib.vinz-clortho'
   project.version = scmVersion.version
-  sourceCompatibility = '11'
+  sourceCompatibility = '17'
 
   repositories {
     if (project.hasProperty('jenkins-ci')) {
@@ -75,7 +75,6 @@ subprojects {
     mavenCentral()
     maven { url "https://s01.oss.sonatype.org/content/repositories/snapshots/" }
     maven { url "https://repo.spring.io/milestone" }
-    maven { url "https://archiva.cogni.zone/repository/public" }
   }
 
   dependencyManagement {
@@ -162,16 +161,6 @@ subprojects {
       }
     }
     repositories {
-      if (project.hasProperty("publishToCognizoneArchiva")) {
-        maven {
-          credentials {
-            username = System.properties['archiva.username']
-            password = System.properties['archiva.password']
-          }
-          name = 'cognizone'
-          url = System.properties['archiva.url']
-        }
-      }
       if (project.hasProperty("publishToCognizoneNexus")) {
         maven {
           credentials {
